add product name in main screen for mobile
