+++ src/pages/questions/ProductInformation.jsx
@@ -29,155 +29,27 @@ import {
   getIsDevMode,
 } from "../../localeStorageManager";
 import { CORRECT_INSIGHT, WRONG_INSIGHT } from "../../const";
-import externalApi from "../../externalApi";
 import DebugQuestion from "./DebugQuestion";
 import ZoomableImage from "../../components/ZoomableImage";
 import robotoff from "../../robotoff";
+import {
+  ADDITIONAL_INFO_TRANSLATION,
+  getImageId,
+  getImagesUrls,
+  useOtherQuestions,
+  useFlagImage,
+} from "./utils";
 
-const ADDITIONAL_INFO_TRANSLATION = {
-  brands: "brands",
-  ingredientsText: "ingredients",
-  countriesTags: "countries",
-  categories: "categories",
-  labels_tags: "labels",
-  quantity: "quantity",
-};
-
-// src looks like: "https://static.openfoodfacts.org/images/products/004/900/053/2258/1.jpg"
-const getImageId = (src) => {
-  const file = src.split("/").at(-1);
-  const imageId = file.replace(/\..+$/, "");
-  return Number(imageId);
-};
-
-const getImagesUrls = (images, barcode) => {
-  if (!images || !barcode) {
-    return [];
-  }
-  const formattedCode = offService.getFormatedBarcode(barcode);
-  const rootImageUrl = offService.getImageUrl(formattedCode);
-  return Object.keys(images)
-    .filter((key) => !isNaN(key))
-    .map((key) => `${rootImageUrl}/${key}.jpg`);
-};
-
-// Other questions fetching
-const useOtherQuestions = (code, insight_id) => {
-  const [otherQuestionsState, setOtherQuestionsState] = React.useState({
-    questions: [],
-    isLoading: true,
-  });
-  React.useEffect(() => {
-    if (!code) {
-      return;
-    }
-    let isStillValid = true;
-    setOtherQuestionsState({
-      questions: [],
-      isLoading: true,
-    });
-    robotoff.questionsByProductCode(code).then((result) => {
-      if (!isStillValid) {
-        return;
-      }
-      const newQuestions = result?.data?.questions ?? [];
-      setOtherQuestionsState({
-        questions: newQuestions.filter((q) => q?.insight_id !== insight_id),
-        isLoading: false,
-      });
-    });
-    return () => {
-      isStillValid = false;
-    };
-  }, [code, insight_id]);
-
-  const [pendingAnswers, setPendingAnswers] = React.useState({});
-  return [
-    otherQuestionsState,
-    setOtherQuestionsState,
-    pendingAnswers,
-    setPendingAnswers,
-  ];
-};
-
-const useFlagImage = (barcode) => {
-  const [flagged, setFlagged] = React.useState([]);
-
-  const flagImage = React.useCallback(
-    (src) => {
-      const imgid = getImageId(src);
-      externalApi.addImageFlag({ barcode, imgid, url: src });
-      setFlagged((prev) => [...prev, imgid]);
-    },
-    [barcode]
-  );
-
-  const deleteFlagImage = React.useCallback(
-    (src) => {
-      const imgid = getImageId(src);
-      externalApi.removeImageFlag({ barcode, imgid });
-
-      setFlagged((prev) =>
-        prev.filter((flaggedImageId) => flaggedImageId !== imgid)
-      );
-    },
-    [barcode]
-  );
-
-  // Reset flags
-  React.useEffect(() => {
-    setFlagged([]);
-  }, [barcode]);
-
-  return [flagged, flagImage, deleteFlagImage];
-};
-
-const ProductInformation = ({ question }) => {
+const ProductInformation = (props) => {
+  const { question, productData } = props;
   const { t } = useTranslation();
   const isDevMode = getIsDevMode();
 
-  const [productData, setProductData] = React.useState({});
   const [hideImages, setHideImages] = React.useState(getHideImages);
   const [devCustomization] = React.useState(
     () => getPageCustomization().questionPage
   );
   const [flagged, flagImage, deleteFlagImage] = useFlagImage(question?.barcode);
-
-  // product data fetching
-  React.useEffect(() => {
-    if (!question?.barcode) {
-      return;
-    }
-    let isStillValid = true;
-    setProductData({
-      code: question.barcode,
-      isLoading: true,
-    });
-    offService.getProduct(question.barcode).then((result) => {
-      if (!isStillValid) {
-        return;
-      }
-      const product = result.data.product;
-      console.log(result.data.product);
-      setProductData({
-        code: question.barcode,
-        productName: product?.product_name || "",
-        brands: product?.brands || "?",
-        ingredientsText: product?.ingredients_text || "?",
-        countriesTags: product?.countries_tags
-          ? `${product?.countries_tags?.join?.(", ")}.`
-          : "?",
-        images: product?.images || {},
-        categories: product?.categories || "?",
-        labels_tags: product?.labels_tags?.join?.(", ") || "?",
-        quantity: product?.quantity || "?",
-        isLoading: false,
-      });
-    });
-    return () => {
-      isStillValid = false;
-    };
-  }, [question?.barcode]);
 
   const [
     otherQuestionsState,
